feat: network and api package
